requirements.txt
streamlit
pandas
supabase
postmarker
python-dotenv